Add circle animation
When I tried to configure the top/bottom/left/right LEDs I mixed
up things a bit. When trying to debug this, it would have helped
to have some plain simple animation that uses these logical LED
arrays to go round the lamp.

So I created the animation "Circle" that draws a red dot moving
round the lamp, starting from bottom left of the lamp, moving
to the right.