Add development helper script
It updates the website from `Website.html`, compiles the firmware
and uploads it to the board via OTA upload.

The script can be called like this:

> compile_and_upload.py [IP-ADDRESS-OF-LAMP]

The script requires Python 3 with the requests module installed.
Additionally it requires the arduino IDE to be in PATH as "arduino"
command.

It has been developed and tested on Linux (Ubuntu). It should work
on MacOS or Windows either out of the box or with smaller changes.