Always start modes with initial settings
In previous versions the board kept runtime information per mode when
switching between modes. This could have odd effects, for example the
"color wipe" mode could be invisible for some time when it was in the
"fade out" phase when switching to the mode.

This change now separates the mode member variables into two categories:

Config: These are initialized with their default values at class level
once and need to be overridden with the applyConfig() method.

State: These are not allowed to be persisted with the config and may
change at any time during animation. They need to be initialized in the
initialize() method.